Safelist distributed web schemes for "registerProtocolHandler"
This CL adds "cabal", "dat", "did", "dweb", "ethereum", "hyper",
"ipfs", "ipns", "ssb" to the safelist of
navigator.registerProtocolHandler. Chrome status entry is [1]
and feature has been discussed in [2] [3] [4].

Currently, the WHATWG and WPT changes are pending on the Mozilla
position review [6]. Web platform exposure is protected by
a new ProtocolHandlerDecentralizedWebSchemes experimental runtime
flag. The validation on the Chrome process is done unconditionally
since that code is not allowed to access Blink's runtime flags and
it is really a browser-side validation to prevent security issue
like [7].

[1] https://www.chromestatus.com/feature/4776602869170176
[2] https://groups.google.com/a/chromium.org/forum/#!msg/blink-dev/29sFh4tTdcs/K4XroilVBAAJ
[3] https://groups.google.com/a/chromium.org/forum/#!topic/blink-dev/7nHTRUP1EGY
[4] https://github.com/whatwg/html/issues/3935
[5] https://github.com/whatwg/html/issues/3998
[6] https://github.com/mozilla/standards-positions/issues/339
[7] crbug.com/971917

Bug: 651311
Change-Id: Iba45706e985015cf86bd80adef990abd0980a638