[scroll-animations] Handle changes to selector(#id) references
An @scroll-timeline rule may reference elements with selector(#id)
syntax. Whenever the element represented by the selector() function
changes, the "effective" ScrollTimeline produced by the
@scroll-timeline rule changes as well.

This CL solves this problem as follows:

 - Maintain a set of all IDs the timelines of a given document
   depend on.
 - When ID-affecting DOM mutations happen, we check if there's a
   match in the set of ID dependencies. If there is, then:
 - We go through all ScrollTimelines associated with the document
   and mark all the animation target elements for style recalc.

Marking the animation target elements for style recalc
(non-animation-style-change) means we'll re-evaluate the
@scroll-timeline rule against the current DOM state, and consider
whether or not we need a new ScrollTimeline for the animation.

Note that we may over-mark with this approach: if there are 100
ScrollTimelines, and some DOM mutation happens which affects *one* of
those, then we still mark *all* associated animation target elements
for style recalc, even though we only really needed to mark a single
element. It's possible to optimize further, but it would require
storing more information. I suggest that we start with the
invalidate-all-timelines approach, and then improve incrementally if
necessary.

It was necessary to add a StyleRuleScrollTimeline pointer to
NewCSSAnimation, UpdateCSSAnimation, and RunningAnimation. This
is because two different @scroll-timeline rules may yield the same
result from ComputeTimeline when both rules reference (two different)
elements that don't exist. We don't need to update the actual
Animation in this case, but we *do* need to update the set of ID
dependencies with the new information as well as the RunningAnimation.

Bug: 1074052
Change-Id: Iab0c6b8d8b57d7d63283e97355ee5d9948b831a0