Perpare to fix cancellign when unsetting transition and fix metric
Add logic to distinguish between transitions which should (base
computed style changes) and shouldn't be cancelled when transition is
completely unset (as the default is all 0s). Currently, both are still
cancelled, the same as the old behaviour, but the histogram
kCSSTransitionCancelledByRemovingStyle is now only incremented when
transitions are cancelled incorrectly.

Add WPT for not cancelling transitions then unsetting transition
property, fix web_tests to cancel transitions by setting transition to
none instead of unsetting.

Bug: 934700
Change-Id: Iaf4bf324e6e2bd6ca22a66fd8ae2320620a874dc