ScrollAnchoring: Limit focus priority to editable elements.
This patch limits the selection of priority candidate to focused
elements with editable style. That is, previously we would consider
any focused element as a priority scroll anchor. Now, this is only
true if the element is editable.

This should be a middle ground that allows most cases in referenced
bug to be fixed (> 30 stars, multiple merged bugs). However, it would
also preserve the layout stability when editing forms with text
input.

R=chrishtr@chromium.org

Fixed: 1102229
Change-Id: I7341f0e513b70f8a4d836cebb68b931f2518534f